wireless-regdb: Update regulatory info for Morocco (MA) on 6GHz
The publication of the Decision ANRT/DG/N°07/2021 of May 2021 amending
and supplementing the technical conditions for the use of radio
installations composed of low power and short range devices [2], which
ANNEX 5 defines p.i.r.e. conditions of 5925 - 6425 MHz are 200 mW and 25 mW
for indoor only and indoor/outdoor respectively.

[1] https://www.anrt.ma/en/lagence/actualites/wifi-6e-now-authorized-morocco
[2] https://www.anrt.ma/sites/default/files/decision_a2fp_-vf-_mod_07.05.2021.pdf?csrt=14818568393101165775

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>